server/workout_app/requirements.txt
asgiref==3.3.4
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2
django-filter==2.4.0
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
idna==2.10
jwcrypto==0.8
Markdown==3.3.4
mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
